+++ Kernel/System/JSON.pm
@@ -66,6 +66,7 @@ Encode a perl data structure to a JSON string.
     my $JSONString = $JSONObject->Encode(
         Data     => $Data,
         SortKeys => 1,          # (optional) (0|1) default 0, to sort the keys of the json data
+        Pretty => 1,            # (optional) (0|1) default 0, to pretty print
     );
 
 =cut
@@ -89,7 +90,12 @@ sub Encode {
 
     # sort the keys of the JSON data
     if ( $Param{SortKeys} ) {
-        $JSONObject->canonical( [1] );
+        $JSONObject->canonical(1);
+    }
+
+    # pretty print - can be useful for debugging purposes
+    if ( $Param{Pretty} ) {
+        $JSONObject->pretty(1);
     }
 
     # get JSON-encoded presentation of perl structure