+++ scripts/test/JSON.t
@@ -83,14 +83,66 @@ my @Tests = (
         Input  => "Some Text with Unicode Characters that  are not allowed\x{2029} in JavaScript",
         Result => '"Some Text with Unicode Characters that\u2028 are not allowed\u2029 in JavaScript"',
         Name   => 'JSON - Unicode Line Terminators are not allowed in JavaScript',
-    }
+    },
+    {
+        Input => [
+            [ 1, 2, "Foo", "Bar" ],
+            {
+                Key1 => 'Something',
+                Key2 => [ "Foo", "Bar" ],
+                Key3 => {
+                    Foo => 'Bar',
+                },
+                Key4 => {
+                    Bar => [ "f", "o", "o" ],
+                }
+            },
+        ],
+        Params => {
+            Pretty => 1,
+        },
+        Result =>
+            '[
+   [
+      1,
+      2,
+      "Foo",
+      "Bar"
+   ],
+   {
+      "Key1" : "Something",
+      "Key2" : [
+         "Foo",
+         "Bar"
+      ],
+      "Key3" : {
+         "Foo" : "Bar"
+      },
+      "Key4" : {
+         "Bar" : [
+            "f",
+            "o",
+            "o"
+         ]
+      }
+   }
+]
+',
+        Name => 'JSON - complex structure - pretty print'
+    },
 );
 
 for my $Test (@Tests) {
 
+    my %Params;
+    if ( $Test->{Params} ) {
+        %Params = %{ $Test->{Params} };
+    }
+
     my $JSON = $JSONObject->Encode(
         Data     => $Test->{Input},
         SortKeys => 1,
+        %Params,
     );
 
     $Self->Is(